Refactors map to its own component
Changed scootertile and mapview to use new mapcomponent


Deleted unused code

Cleanup

Fixed issues from rebase
