+++ src/components/Map/BikeRentalStationTag/index.tsx
