+++ src/components/Map/BikeRentalStationTag/styles.scss
