+++ src/components/Map/StopPlaceTag/styles.scss
