+++ src/dashboards/Compact/styles.scss
@@ -17,7 +17,7 @@
 
         .scootertile {
             width: 100%;
-            overflow: hidden;
+            padding-top: 2rem;
             margin-bottom: 1rem;
         }
     }
