--- src/dashboards/Map/MapView/styles.scss
@@ -1,12 +1,0 @@
-@import '../../../assets/styles/import.scss';
-@import '../../../variables.scss';
-
-.departure-display {
-    width: 100%;
-    display: flex;
-    position: absolute;
-    height: auto;
-    bottom: 1.5rem;
-    align-items: center;
-    justify-content: center;
-}
