+++ src/dashboards/Map/index.tsx
@@ -7,14 +7,20 @@ import {
     useWalkTime,
 } from '../../logic'
 
+import MapView from '../../components/Map'
+
+import DepartureTag from './DepartureTag'
 import './styles.scss'
-import MapView from './MapView'
 
 const MapDashboard = ({ history }: Props): JSX.Element => {
     const stopPlacesWithDepartures = useStopPlacesWithDepartures()
     const bikeRentalStations = useBikeRentalStations()
     const walkTimes = useWalkTime(stopPlacesWithDepartures)
-
+    const HEADER_MARGIN = 16
+    //Used to calculate the height of the viewport for the map
+    const headerHeight =
+        (document?.getElementsByClassName('header')[0]?.clientHeight ?? 0) +
+        HEADER_MARGIN
     return (
         <DashboardWrapper
             className="map-tile"
@@ -22,11 +28,27 @@ const MapDashboard = ({ history }: Props): JSX.Element => {
             stopPlacesWithDepartures={stopPlacesWithDepartures}
             bikeRentalStations={bikeRentalStations}
         >
-            <MapView
-                bikeRentalStations={bikeRentalStations}
-                stopPlacesWithDepartures={stopPlacesWithDepartures}
-                walkTimes={walkTimes}
-            ></MapView>
+            <div style={{ height: `calc(100vh - ${headerHeight}px)` }}>
+                <MapView
+                    scooters={null}
+                    bikeRentalStations={bikeRentalStations}
+                    stopPlaces={stopPlacesWithDepartures}
+                    walkTimes={walkTimes}
+                    interactable={true}
+                ></MapView>
+                <div className="departure-display">
+                    {stopPlacesWithDepartures?.map((sp) =>
+                        sp.departures.length ? (
+                            <DepartureTag
+                                key={sp.id}
+                                stopPlace={sp}
+                            ></DepartureTag>
+                        ) : (
+                            []
+                        ),
+                    )}
+                </div>
+            </div>
         </DashboardWrapper>
     )
 }
