+++ src/dashboards/Map/styles.scss
@@ -3,8 +3,16 @@
 
 .map-tile {
     width: 100%;
-    float: left;
-    box-sizing: border-box;
+    color: white;
     padding: 0;
-    position: relative;
+}
+
+.departure-display {
+    width: 100%;
+    display: flex;
+    position: absolute;
+    height: auto;
+    bottom: 1.5rem;
+    align-items: center;
+    justify-content: center;
 }
